// src/models/romanEnum.model.ts
/* eslint-disable prettier/prettier */
export enum RomanEnum {
	I = 'I',
	II = 'II',
	III = 'III',
	IV 	= 'IV',
	V = 'V',
	VI = 'VI',
	VII = 'VII',
	VIII = 'VIII',
	IX = 'IX',
	X = 'X',
	XX = 'XX',
	XXX = 'XXX',
	XL = 'XL',
	L = 'L',
	LX = 'LX',
	LXX = 'LXX',
	LXXX = 'LXXX',
	XC = 'XC',
	C = 'C',
	CC = 'CC',
	CCC = 'CCC',
	CD = 'CD',
	D = 'D',
	DC = 'DC',
	DCC = 'DCC',
	DCCC = 'DCCC',
	CM = 'CM',
	M = 'M',
	MM = 'MM',
	MMM = 'MMM'
}